+++ elina_poly/opt_pk_project.c
@@ -378,8 +378,8 @@ void opt_poly_projectforget_array(bool project,
 	}
 	//quasi_removal(opk,op);
 	nbcons = ocp->nbrows;
-        int limit = nbcons;
-        int  nbadd;
+	size_t limit = nbcons;	
+	int  nbadd;
 	int maxadd;
 	while(size > 0){
